move GetProof to ipld pkg
