+++ share/ipld/get.go
@@ -157,46 +157,6 @@ func GetLeaves(ctx context.Context,
 	wg.Wait()
 }
 
-// GetProof fetches and returns the leaf's Merkle Proof.
-// It walks down the IPLD NMT tree until it reaches the leaf and returns collected proof
-func GetProof(
-	ctx context.Context,
-	bGetter blockservice.BlockGetter,
-	root cid.Cid,
-	proof []cid.Cid,
-	leaf, total int,
-) ([]cid.Cid, error) {
-	// request the node
-	nd, err := GetNode(ctx, bGetter, root)
-	if err != nil {
-		return nil, err
-	}
-	// look for links
-	lnks := nd.Links()
-	if len(lnks) == 0 {
-		p := make([]cid.Cid, len(proof))
-		copy(p, proof)
-		return p, nil
-	}
-
-	// route walk to appropriate children
-	total /= 2 // as we are using binary tree, every step decreases total leaves in a half
-	if leaf < total {
-		root = lnks[0].Cid // if target leave on the left, go with walk down the first children
-		proof = append(proof, lnks[1].Cid)
-	} else {
-		root, leaf = lnks[1].Cid, leaf-total // otherwise go down the second
-		proof, err = GetProof(ctx, bGetter, root, proof, leaf, total)
-		if err != nil {
-			return nil, err
-		}
-		return append(proof, lnks[0].Cid), nil
-	}
-
-	// recursively walk down through selected children
-	return GetProof(ctx, bGetter, root, proof, leaf, total)
-}
-
 // chanGroup implements an atomic wait group, closing a jobs chan
 // when fully done.
 type chanGroup struct {
